rename `replaceCustomEmoji.js` to `customEmoji.js`
